Code clean-up, moved users to separate python module
