+++ skeleton/admin.py
@@ -1,7 +1,3 @@
-from django.contrib import admin
-
-from skeleton.models import UserModel
-
 # Register your models here.
 from django.contrib import admin
 from skeleton.models import UserModel
