+++ skeleton/models.py
@@ -1,6 +1,2 @@
 # Create your models here.
-from django.db import models
-
-class UserModel(models.Model):
-    name = models.CharField(max_length=100)
-    last_name = models.CharField(max_length=100)
+from .users.model import UserModel
