+++ skeleton/users/model.py
@@ -1,0 +1,5 @@
+from django.db import models
+
+class UserModel(models.Model):
+    name = models.CharField(max_length=100)
+    last_name = models.CharField(max_length=100)
